Update and make more user friendly to README.md
I have make README.md more attractive and added emoji. The plus point is that I added link on data.Brno, as it will help user to know about data.Brno if they don't know about it. I also added more content in it. 

if you like my work so accept this because it will motivate.